@@ -4,25 +4,28 @@ import { provideRouter } from '@angular/router';
 import { routes } from './app.routes';
 import { provideClientHydration } from '@angular/platform-browser';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
-import { provideHttpClient, withFetch } from '@angular/common/http';
+import {
+  provideHttpClient,
+  withFetch,
+  withInterceptors,
+  withInterceptorsFromDi,
+} from '@angular/common/http';
 import { provideToastr } from 'ngx-toastr';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { authConfig } from './auth/auth.config';
-import { provideAuth } from 'angular-auth-oidc-client';
-
-
+import { authInterceptor, provideAuth } from 'angular-auth-oidc-client';
 
 export const appConfig: ApplicationConfig = {
-
   providers: [
-    provideAuth(authConfig),
-    provideRouter(routes),
-    provideClientHydration(),
-    provideAnimationsAsync(),
-    provideHttpClient(withFetch()),
     provideAnimations(),
+    provideAnimationsAsync(),
+    provideHttpClient(withFetch(), withInterceptors([authInterceptor()])),
+    provideAuth(authConfig),
+    provideClientHydration(),
+    provideRouter(routes),
     provideToastr({
-      timeOut: 5000
-    })
-  ]
+      timeOut: 4000,
+      preventDuplicates: true,
+    }),
+  ],
 };
