@@ -8,5 +8,7 @@ export interface VideoDetails {
     videoUrl: string;
     thumbnailUrl: string;
     likeCount?: number;
-    dislikeCount?: number;
+    disLikeCount?: number;
+    viewCount?: number;
+    createdAt:string
 }
