@@ -6,60 +6,62 @@ import { MatInputModule } from '@angular/material/input';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { FlexLayoutServerModule } from '@angular/flex-layout/server';
 import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
-import { MatChipEditedEvent, MatChipInputEvent, MatChipsModule } from '@angular/material/chips';
+import {
+  MatChipEditedEvent,
+  MatChipInputEvent,
+  MatChipsModule,
+} from '@angular/material/chips';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { LiveAnnouncer } from '@angular/cdk/a11y';
 import { UploadThumbnailComponent } from './upload-thumbnail/upload-thumbnail.component';
 import { VideoService } from '../services/video.service';
-import { VideoPlayerComponent } from "../video-player/video-player.component";
-import { Subscription, timeout } from 'rxjs';
-
+import { VideoPlayerComponent } from '../video-player/video-player.component';
+import { Subscription } from 'rxjs';
 
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { CommonModule } from '@angular/common';
 import { changeLoaderStatus } from '../shared/shared-function';
 import { VideoDetails } from '../interfaces/video-details';
 import { ToastrModule, ToastrService } from 'ngx-toastr';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
+import { WAIT_TIME } from '../shared/system.properties';
 
 @Component({
-  selector: 'app-save-video-details',
-  standalone: true,
-  imports: [
-    CommonModule,
-    ReactiveFormsModule,
-    FlexLayoutModule,
-    FlexLayoutServerModule,
-    MatFormFieldModule,
-    MatInputModule,
-    MatSelectModule,
-    MatButtonModule,
-    MatChipsModule,
-    MatIconModule,
-    UploadThumbnailComponent,
-    VideoPlayerComponent,
-    MatProgressSpinnerModule,
-    ToastrModule
-  ],
-  templateUrl: './save-video-details.component.html',
-  styleUrl: './save-video-details.component.scss'
+    selector: 'app-save-video-details',
+    standalone: true,
+    imports: [
+        CommonModule,
+        ReactiveFormsModule,
+        FlexLayoutModule,
+        FlexLayoutServerModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatSelectModule,
+        MatButtonModule,
+        MatChipsModule,
+        MatIconModule,
+        UploadThumbnailComponent,
+        VideoPlayerComponent,
+        MatProgressSpinnerModule,
+        ToastrModule,
+    ],
+    templateUrl: './save-video-details.component.html',
+    styleUrl: './save-video-details.component.scss'
 })
 export class SaveVideoDetailsComponent implements OnInit, OnDestroy {
-
-
   //subscriptions
-  private checkThumbnailStatus$!: Subscription
-  private getVideoDetail$!: Subscription
+  private checkThumbnailStatus$!: Subscription;
+  private getVideoDetail$!: Subscription;
   loading: boolean = true;
   //form properties
   saveVideoDetailForm: FormGroup;
   title: FormControl = new FormControl('');
   description: FormControl = new FormControl('');
   videoStatus: FormControl = new FormControl('');
-  fileSelected: boolean = false
+  fileSelected: boolean = false;
   //video variables
-  videoId: string = '';
+  videoId!: string | '';
   thumbnailUrl: string = '';
   videoUrl!: string;
   readonly addOnBlur = true;
@@ -70,47 +72,58 @@ export class SaveVideoDetailsComponent implements OnInit, OnDestroy {
   //btn check status
   btnDisabled: boolean = false;
 
-
   constructor(
     private activatedRoute: ActivatedRoute,
+    private router: Router,
     public videoService: VideoService,
     private toastr: ToastrService
   ) {
     // get videoId
     this.videoId = this.activatedRoute.snapshot.params['videoId'];
-    
+
     //create form group
     this.saveVideoDetailForm = new FormGroup({
       title: this.title,
       description: this.description,
-      videoStatus: this.videoStatus
-    })
-
-  };
+      videoStatus: this.videoStatus,
+    });
+  }
 
   ngOnInit(): void {
     // check upload thumbnail status
-    this.checkThumbnailStatus$ = this.videoService.uploadThumbnailStatus$.subscribe(
-      status => {
-        this.fileSelected = status
-      }
-    )
+    this.checkThumbnailStatus$ =
+      this.videoService.uploadThumbnailStatus$.subscribe((status) => {
+        this.fileSelected = status;
+      });
     //get video details
-    this.getVideoDetail$ = this.videoService.getVideoDetails(this.videoId).subscribe(
-      videoDetails => {
-        this.title.patchValue(videoDetails.title)
-        this.description.patchValue(videoDetails.description)
-        this.videoStatus.patchValue(videoDetails.videoStatus)
-        this.videoUrl = videoDetails.videoUrl
-        this.thumbnailUrl = videoDetails.thumbnailUrl
-        this.tags.set(this.tags().concat(videoDetails.tags))
-      }
-    )
+    if (this.videoId !== '') {
+      this.getVideoDetail$ = this.videoService
+        .getVideoDetails(this.videoId)
+        .subscribe({
+          next: (videoDetails) => {
+            this.title.patchValue(videoDetails.title);
+            this.description.patchValue(videoDetails.description);
+            this.videoStatus.patchValue(videoDetails.videoStatus);
+            this.videoUrl = videoDetails.videoUrl;
+            this.thumbnailUrl = videoDetails.thumbnailUrl;
+            this.tags.set(this.tags().concat(videoDetails.tags));
+          },
+          error: (e) => {
+            this.toastr.error(e.statusText, e.status, {
+              timeOut: WAIT_TIME,
+            });
+            if ((e.status = '401')) {
+              this.router.navigate(['upload-video']);
+            }
+          },
+          complete: () => {
+            changeLoaderStatus().then((status) => {
+              this.loading = status;
+            });
+          },
+        });
+    }
     //
-    changeLoaderStatus().then(status => {
-      this.loading = status
-    })
-
   }
 
   add(event: MatChipInputEvent): void {
@@ -118,7 +131,7 @@ export class SaveVideoDetailsComponent implements OnInit, OnDestroy {
 
     // Add our fruit
     if (value) {
-      this.tags.update(tags => [...tags, value]);
+      this.tags.update((tags) => [...tags, value]);
     }
 
     // Clear the input value
@@ -126,7 +139,7 @@ export class SaveVideoDetailsComponent implements OnInit, OnDestroy {
   }
 
   remove(Tags: string): void {
-    this.tags.update(tags => {
+    this.tags.update((tags) => {
       const index = tags.indexOf(Tags);
       if (index < 0) {
         return tags;
@@ -148,7 +161,7 @@ export class SaveVideoDetailsComponent implements OnInit, OnDestroy {
     }
 
     // Edit existing fruit
-    this.tags.update(tags => {
+    this.tags.update((tags) => {
       const index = tags.indexOf(Tags);
       if (index >= 0) {
         tags[index] = value;
@@ -158,6 +171,8 @@ export class SaveVideoDetailsComponent implements OnInit, OnDestroy {
     });
   }
   saveVideo() {
+    const currentDate = new Date();
+
     this.btnDisabled = true;
     const videoDetails: VideoDetails = {
       videoId: this.videoId,
@@ -167,25 +182,24 @@ export class SaveVideoDetailsComponent implements OnInit, OnDestroy {
       videoStatus: this.videoStatus.value,
       tags: this.tags(),
       videoUrl: this.videoUrl,
-      thumbnailUrl: this.thumbnailUrl
-    }
+      thumbnailUrl: this.thumbnailUrl,
+      createdAt: currentDate.toISOString()
+    };
 
     this.videoService.saveVideoDetails(videoDetails).subscribe(
-      response => {
+      (response) => {},
+      (error) => {
+        console.log('1');
 
-      },
-      error => {
-        this.toastr.error(error.message)
+        this.toastr.error(error.message);
         this.btnDisabled = false;
       },
       () => {
-        this.toastr.success('Video details saved successfully')
+        this.toastr.success('Video details saved successfully');
         this.btnDisabled = false;
       }
-    )
-
+    );
   }
-
 
   //unsubscribing
   ngOnDestroy(): void {
@@ -195,8 +209,5 @@ export class SaveVideoDetailsComponent implements OnInit, OnDestroy {
     if (this.getVideoDetail$) {
       this.getVideoDetail$.unsubscribe();
     }
-
-
   }
-
 }
