@@ -1,2 +1,2 @@
-export const TIME_OUT: number = 3000
+export const TIME_OUT: number = 1000
 export const WAIT_TIME: number = TIME_OUT - 500
